+++ sdk/keyvault/azure-mgmt-keyvault/azure/mgmt/keyvault/v2016_10_01/models/_models.py
@@ -587,7 +587,7 @@ class VaultCheckNameAvailabilityParameters(msrest.serialization.Model):
 
     :param name: Required. The vault name.
     :type name: str
-    :ivar type: The type of resource, Microsoft.KeyVault/vaults. Has constant value:
+    :ivar type: Required. The type of resource, Microsoft.KeyVault/vaults. Default value:
      "Microsoft.KeyVault/vaults".
     :vartype type: str
     """