+++ sdk/keyvault/azure-mgmt-keyvault/azure/mgmt/keyvault/v2016_10_01/operations/_vaults_operations.py
@@ -710,8 +710,8 @@ class VaultsOperations(object):
         :type location: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: By default, your polling method will be ARMPolling.
-         Pass in False for this operation to not poll, or pass in your own initialized polling object for a personal polling strategy.
+        :keyword polling: Pass in True if you'd like the ARMPolling polling method,
+         False for no polling, or your own initialized polling object for a personal polling strategy.
         :paramtype polling: bool or ~azure.core.polling.PollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of LROPoller that returns either None or the result of cls(response)
@@ -762,12 +762,18 @@ class VaultsOperations(object):
 
     def list(
         self,
+        filter,  # type: Union[str, "_models.Enum5"]
+        api_version,  # type: Union[str, "_models.Enum6"]
         top=None,  # type: Optional[int]
         **kwargs  # type: Any
     ):
         # type: (...) -> Iterable["_models.ResourceListResult"]
         """The List operation gets information about the vaults associated with the subscription.
 
+        :param filter: The filter to apply on the operation.
+        :type filter: str or ~azure.mgmt.keyvault.v2016_10_01.models.Enum5
+        :param api_version: Azure Resource Manager Api Version.
+        :type api_version: str or ~azure.mgmt.keyvault.v2016_10_01.models.Enum6
         :param top: Maximum number of results to return.
         :type top: int
         :keyword callable cls: A custom type or function that will be passed the direct response
@@ -780,8 +786,6 @@ class VaultsOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        filter = "resourceType eq 'Microsoft.KeyVault/vaults'"
-        api_version = "2015-11-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):