+++ sdk/keyvault/azure-mgmt-keyvault/azure/mgmt/keyvault/v2021_06_01_preview/aio/operations/_private_link_resources_operations.py
@@ -44,7 +44,7 @@ class PrivateLinkResourcesOperations:
         self,
         resource_group_name: str,
         vault_name: str,
-        **kwargs: Any
+        **kwargs
     ) -> "_models.PrivateLinkResourceListResult":
         """Gets the private link resources supported for the key vault.