+++ sdk/keyvault/azure-mgmt-keyvault/azure/mgmt/keyvault/v2021_06_01_preview/operations/_vaults_operations.py
@@ -123,8 +123,8 @@ class VaultsOperations(object):
         :type parameters: ~azure.mgmt.keyvault.v2021_06_01_preview.models.VaultCreateOrUpdateParameters
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: By default, your polling method will be ARMPolling.
-         Pass in False for this operation to not poll, or pass in your own initialized polling object for a personal polling strategy.
+        :keyword polling: Pass in True if you'd like the ARMPolling polling method,
+         False for no polling, or your own initialized polling object for a personal polling strategy.
         :paramtype polling: bool or ~azure.core.polling.PollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of LROPoller that returns either Vault or the result of cls(response)
@@ -773,8 +773,8 @@ class VaultsOperations(object):
         :type location: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: By default, your polling method will be ARMPolling.
-         Pass in False for this operation to not poll, or pass in your own initialized polling object for a personal polling strategy.
+        :keyword polling: Pass in True if you'd like the ARMPolling polling method,
+         False for no polling, or your own initialized polling object for a personal polling strategy.
         :paramtype polling: bool or ~azure.core.polling.PollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of LROPoller that returns either None or the result of cls(response)
@@ -825,12 +825,18 @@ class VaultsOperations(object):
 
     def list(
         self,
+        filter,  # type: Union[str, "_models.Enum16"]
+        api_version,  # type: Union[str, "_models.Enum17"]
         top=None,  # type: Optional[int]
         **kwargs  # type: Any
     ):
         # type: (...) -> Iterable["_models.ResourceListResult"]
         """The List operation gets information about the vaults associated with the subscription.
 
+        :param filter: The filter to apply on the operation.
+        :type filter: str or ~azure.mgmt.keyvault.v2021_06_01_preview.models.Enum16
+        :param api_version: Azure Resource Manager Api Version.
+        :type api_version: str or ~azure.mgmt.keyvault.v2021_06_01_preview.models.Enum17
         :param top: Maximum number of results to return.
         :type top: int
         :keyword callable cls: A custom type or function that will be passed the direct response
@@ -843,8 +849,6 @@ class VaultsOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        filter = "resourceType eq 'Microsoft.KeyVault/vaults'"
-        api_version = "2015-11-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):